brain_model_train/requirements.txt
monai[nibabel]
pytorch-lightning
wandb
omegaconf
hydra-core
einops
pillow
boto3